Move callbacks to separate file
